[#29] ProgressState 파일분리, 프로그래스바 구현 완료
